Introduce configurable access modifiers for generated code
This commit extends the YAML configuration syntax with an option to
specify the access modifiers for generated code with secret literal
declarations, both globally and on per secret basis.

The supported access modifiers are `internal` (default) and `public`.
The use of this option is optional, and if not specified, the default
behavior is the same as before the change. With that said, this is a
non-breaking change.

Example usage:
- `defaultAccessModifier: public` -> change global access modifier to
  `public`. This option allows you to specify the default access
  modifier applied to each secret, unless the secret definition states
  otherwise.
- `accessModifier: internal` -> used at secret definition level,
  specifies that this secret literal shall be declared `internal`
  regardless of the value of `defaultAccessModifier` option.