Add support for creating XCFrameworks that depend on ConfidentialKit
This commit extends the YAML configuration syntax with
`implementationOnlyImport` boolean option that can be used to enable
the generation of implementation-only `ConfidentialKit` import.
The use of this option is optional, and if not specified, the default
value is `false`, i.e. this is a non-breaking change.

Includes the removal of `@inlinable` attributes from some initializers
in `ConfidentialKit` module that were causing compiler errors when
building libraries for distribution (i.e. with library evolution and
generation of a module interface file enabled).