Add Swift 6 language mode support
Introduces `@Obfuscated` macro aimed to be a long-term replacement for
`@Obfuscated` property wrapper. Since property wrappers can only be
applied to `var` properties, a static property wrapped in a property
wrapper is not considered concurrency-safe and causes the compiler to
emit an error when complete concurrency checking is enabled (i.e.
Swift 6 language mode). To address this issue, a decision was made to
rewrite the existing `@Obfuscated` property wrapper to a macro.

Alas, due to its noticeable impact on build time and (still) immature
tooling around Swift macros, the `@Obfuscated` macro is only available
as part of the experimental `_ConfidentialKit` library (Swift 6 only).
To make `swift-confidential` use experimental APIs for generated source
code, one must enable the Experimental Mode via the `experimentalMode`
boolean configuration option (defaults to `false`).

Additionally, includes generation of `internal` import to resolve the
"'@_implementationOnly' is deprecated, use 'internal import' instead"
warning emitted by the Swift 6.0 compiler.
The `implementationOnlyImport` configuration option was renamed to
`internalImport`, but the old name is still supported for backward
compatibility. Likewise, the generated code includes the
`#if compiler(>=6.0) || hasFeature(AccessLevelOnImport)` check to
switch between `internal` and `@_implementationOnly` imports to ensure
compatibility with Swift 5.9 and 5.10 compilers that do not support
access-level modifiers on import declarations out of the box.