Introduce random obfuscation algorithm generation
This commit extends the YAML configuration syntax with an option to
enable random obfuscation algorithm generation (i.e. by specifying
`algorithm: random`). This new mode of operation is now the default,
such that explicitly specifying the obfuscation algorithm is no longer
required.

The `SourceFileSpec.ObfuscationStep` data structure was flattened into a
simpler, non-nested construct.

This is a non-breaking change.